[services][checker] move spreadsheet to excel package
