+++ ws/src/main/scala/de/thm/ii/fbs/controller/TaskController.scala
@@ -6,7 +6,7 @@ import java.io.File
 import com.fasterxml.jackson.databind.JsonNode
 import de.thm.ii.fbs.controller.exception.{BadRequestException, ForbiddenException, ResourceNotFoundException}
 import de.thm.ii.fbs.model.{CourseRole, GlobalRole, SpreadsheetMediaInformation, SpreadsheetResponseInformation, SubtaskStatisticsTask, Task, UserTaskResult}
-import de.thm.ii.fbs.services.checker.SpreadsheetService
+import de.thm.ii.fbs.services.checker.excel.SpreadsheetService
 import de.thm.ii.fbs.services.security.AuthService
 import de.thm.ii.fbs.util.Hash
 import de.thm.ii.fbs.util.JsonWrapper.jsonNodeToWrapper
