+++ ws/src/main/scala/de/thm/ii/fbs/services/checker/excel/SpreadsheetCheckerService.scala
@@ -1,14 +1,15 @@
-package de.thm.ii.fbs.services.checker
+package de.thm.ii.fbs.services.checker.excel
 
 import com.fasterxml.jackson.databind.json.JsonMapper
 import de.thm.ii.fbs.model.{CheckrunnerConfiguration, SpreadsheetMediaInformation, User}
+import de.thm.ii.fbs.services.checker.CheckerService
 import de.thm.ii.fbs.services.persistence.{CheckrunnerSubTaskService, StorageService, SubmissionService, TaskService}
 import de.thm.ii.fbs.util.Hash
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.stereotype.Service
 
 import java.io.File
-import java.text.{DecimalFormat, DecimalFormatSymbols, NumberFormat, ParseException}
+import java.text.{NumberFormat, ParseException}
 import java.util.{Locale, Map => UtilMap}
 import scala.collection.mutable
